client/public/base_integration/uiauto_executor/requirements.txt
Pillow==6.2.2
numpy==1.22.3
asyncio==3.4.3
aiohttp==3.8.1
eventlet==0.33.0
JPype1==1.3.0
python-engineio==3.14.2
python-socketio==4.6.0
requests==2.27.1
selenium==3.141.0